refactor: 공공 API 받아오는 부분 분리
